Add functions for login system

+++ bloodTest/src/main/java/com/blood/dao/StaffDAO.java
@@ -3,7 +3,7 @@ package com.blood.dao;
 import com.blood.pojo.Staff;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StaffDAO extends JpaRepository<Staff,Integer>{
+public interface StaffDAO extends JpaRepository<Staff, Long>{
     Staff findByUsername(String username);
     Staff findById(int id);
 }
